+++ instagram/post/models.py
@@ -25,6 +25,7 @@ from sqlalchemy import (
 )
 from sqlalchemy.ext.hybrid import hybrid_method, hybrid_property
 from sqlalchemy.orm import relationship
+from sqlalchemy.sql.expression import and_, true
 
 db = SessionLocal()
 
@@ -62,6 +63,14 @@ class Post(LikeableEntity):
     @hybrid_property
     def comment_count(self):
         return len(self.comments)
+
+    @hybrid_method
+    def is_liked_by(self, user: InstagramUser):
+        return any(like.user_id == user.id for like in self.likes)
+
+    @is_liked_by.expression
+    def is_liked_by(cls, user: InstagramUser):
+        return and_(true(), cls.likes.any(user_id=user.id))
 
     @hybrid_method
     def is_saved_by(self, user: InstagramUser):
